Abstract virtual channel test
This makes the test reusable for different backends.
